Reorganise code layout
