+++ src/providers/esplora.ts
@@ -1,20 +1,31 @@
-import { fetchData, LOGLEVEL } from "./util";
-import { logger } from "./logger";
+import { fetchData, LOGLEVEL } from "../lib/util";
+import { logger } from "../lib/logger";
 
 const log = logger(LOGLEVEL);
 
 /**
- * MempoolProvider class implements the Provider interface.
- * It provides methods to fetch data from a Mempool API.
+ * A class that provides data from an Esplora server.
+ *
+ * The `EsploraProvider` class fetches data such as the current block height, block hash,
+ * and fee estimates from an Esplora server. It provides methods to fetch each of these
+ * data points individually, as well as a method to fetch all of them at once.
+ *
+ * This class implements the `Provider` interface, ensuring it provides all necessary
+ * methods and properties for a data provider.
+ *
+ * @example
+ * const provider = new EsploraProvider('https://blockstream.info/api/');
+ * const data = await provider.getAllData();
  */
-export class MempoolProvider implements Provider {
+export class EsploraProvider implements Provider {
   private url: string;
   private depth: number;
   private timeout: number;
 
   /**
-   * Constructs a new MempoolProvider.
-   * @param url - The base URL of the Mempool API.
+   * Constructs a new EsploraProvider.
+   *
+   * @param url - The base URL of the Esplora API.
    * @param defaultDepth - The default depth for fee estimates.
    * @param defaultTimeout - The default timeout for fetch requests.
    */
@@ -29,22 +40,19 @@ export class MempoolProvider implements Provider {
   }
 
   /**
-   * Fetches fee estimates from the Mempool API.
+   * Fetches fee estimates from the Esplora API.
+   *
    * @param maxDepth - The maximum depth for fee estimates.
    * @returns A promise that resolves to an object of fee estimates.
    */
   async getFeeEstimates(
     maxDepth: number = this.depth,
   ): Promise<FeeByBlockTarget> {
-    const data = await fetchData<MempoolFeeEstimates>(
-      `${this.url}/api/v1/fees/recommended`,
+    const data = await fetchData<EsploraFeeEstimates>(
+      `${this.url}/api/fee-estimates`,
       "json",
       this.timeout,
     );
-
-    if (!data.fastestFee || !data.halfHourFee || !data.hourFee) {
-      throw new Error("Invalid fee data");
-    }
 
     const feeEstimates: FeeByBlockTarget = this.transformFeeData(
       data,
@@ -54,7 +62,8 @@ export class MempoolProvider implements Provider {
   }
 
   /**
-   * Fetches the current block height from the Mempool API.
+   * Fetches the current block height from the Esplora API.
+   *
    * @returns A promise that resolves to the current block height.
    */
   async getBlockHeight(): Promise<number> {
@@ -72,7 +81,8 @@ export class MempoolProvider implements Provider {
   }
 
   /**
-   * Fetches the current block hash from the Mempool API.
+   * Fetches the current block hash from the Esplora API.
+   *
    * @returns A promise that resolves to the current block hash.
    */
   async getBlockHash(): Promise<string> {
@@ -93,7 +103,8 @@ export class MempoolProvider implements Provider {
   }
 
   /**
-   * Fetches all data (block height, block hash, and fee estimates) from the Mempool API.
+   * Fetches all data (block height, block hash, and fee estimates) from the Esplora API.
+   *
    * @returns A promise that resolves to an object of all data.
    */
   public async getAllData(): Promise<ProviderData> {
@@ -110,26 +121,34 @@ export class MempoolProvider implements Provider {
         feeEstimates,
       };
     } catch (error) {
-      log.error({ msg: "Error fetching all data from Mempool:", error });
+      log.error({ msg: "Error fetching all data from Esplora:", error });
       throw error;
     }
   }
 
   /**
    * Transforms the fetched fee data into a FeeByBlockTarget object.
+   *
    * @param data - The fetched fee data.
    * @param maxDepth - The maximum depth for fee estimates.
    * @returns A FeeByBlockTarget object.
    */
   private transformFeeData(
-    data: MempoolFeeEstimates,
+    data: EsploraFeeEstimates,
     maxDepth: number,
   ): FeeByBlockTarget {
     const feeEstimates: FeeByBlockTarget = {};
 
-    if (maxDepth >= 1) feeEstimates[1] = data.fastestFee;
-    if (maxDepth >= 3) feeEstimates[3] = data.halfHourFee;
-    if (maxDepth >= 6) feeEstimates[6] = data.hourFee;
+    // Iterate over the keys in the data object
+    for (const target in data) {
+      // Convert the target to a number
+      const targetNumber = Number(target);
+
+      // If the target is less than or equal to the maximum depth, add it to the fee estimates
+      if (!isNaN(targetNumber) && targetNumber <= maxDepth) {
+        feeEstimates[targetNumber] = data[target];
+      }
+    }
 
     return feeEstimates;
   }
